Refactor: 아이템 보상 지급 수정
아이템 보상지급 int -> enum 수정

+++ 03_etc/ItemQuest.cs
@@ -29,7 +29,7 @@ namespace TeamRPG_17
                 GameManager.Instance.player.AddExp(exp);
                 GameManager.Instance.player.gold += gold;
                 if (rewardItem != null)
-                    GameManager.Instance.player.inventory.AddItem((int)rewardItem);
+                    GameManager.Instance.player.inventory.AddItem(rewardItem);
                 return true;
             }
