+++ 03_etc/KillQuest.cs
@@ -31,7 +31,7 @@ namespace TeamRPG_17
                 GameManager.Instance.player.AddExp(exp);
                 GameManager.Instance.player.gold += gold;
                 if(rewardItem != null)
-                    GameManager.Instance.player.inventory.AddItem((int)rewardItem);
+                    GameManager.Instance.player.inventory.AddItem(rewardItem);
                 return true;
             }
